test(collator): The empty case
I find myself reaching for the Zero-One-Many principle at the beginning of most greenfield TDD exercises. It's a great way to start sketching out the domain of the problem without over-committing.

As a result, the first test I wrote was for the case where we've submitted no test results; in this case we're expecting that the `endRun()` method will return an empty collection of `TestResult`.

This is nice, because we can specify that we want a list of `TestResult` but we don't actually need to specify what that enum contains. Having an empty enum is more than sufficient for Java to happily compile, so this makes a nice first dip into the water.