test(collator): One is enough for now
After zero inevitably comes one. This task has multiple dimensions of complexity; recording tests, recording multiple runs, determining whether a test is new or not.

I elect to ignore all but the "recording tests" part at this stage, mainly because no-one can stop me but also because it is the smallest complexity jump that will also drive the design. Recording multiple runs has no meaning if we can't submit tests, and the same is true for determining whether tests are new or not.

The simplest thing to do here is to create a `TestResult` from the input to `add()` and then return a list of these things. If I was being _really_ strict I would instead store the result locally and add it to the list in the `endRun()` step, but again, no-one can stop me so I decided to collapse a few of the very small steps into one still-small step.

I know that the `TestStatus` can take many values other than PASS, but they aren't relevant at this point in time so I chose not to add them.