test(collator): Nothing new under the sun
At this point I have a fork in the road; which capability do I address next? I chose to tackle whether a test is new or not because I can take a simple step without requiring me to build out the concept of "unrun tests" just yet. If I submit the same single test again after getting the run summary, I should see the same test result but the only difference is that `isNew` is `false` instead of the default `true`.

This requires the component to keep track of which tests it has seen so far, and a `Set` is the right concept for this. I chose a `HashSet` for now but I will likely have to revisit this when I write tests for consistent ordering. For now, though, this will suffice; I don't want to implement too much ahead of time if I can help it.

I need to take a copy of the results list in `endRun()` because if I pass the reference to the consumer and then call `.clear()` the test will have an empty list rather than the original list.