refactor(collator): Refactoring Interlude #1
This was a good place to step back and do a bit of cleaning up. I had a lot of duplication in places which I cleaned up by writing a wrapper around the test matcher. Now the test cases read a lot more fluently and accurately describe what the component is doing.

I'm a big fan of custom matchers in general because they are a great way to clearly express the properties of your tests within your own domain language. I view only using the base matchers the way I view primitive obsession - fine up until a point, but we have language to describe these things better, so let's do it!

After this quick pit-stop in refactoring land, and with fresh fuel in the tank, I decided to pick up on the next bit of functionality.