test(collator): Coming un-run
The next bit of complexity I decided to address was the concept of unrun tests. This means that if we submit a passing test `test1` and then call `endRun()` twice, the second call will return the fact that `test1` was unrun rather than pass or fail.

This is pretty normal behaviour in a lot of IDEs and testing frameworks so I can see why it's in here. I wrote a simple test and in the spirit of the simplest thing that could work, I took the set of tests that I'd seen and looped over which of them were in the most recent list of results. For any that weren't, I'd create a new `TestResult` with a status of `UNRUN`. We can also cheat a bit here; by definition, tests that have been run once but not since are not new, so `isNew` will always be false.

However, the "simplest thing that could work" ends up looking like this:

```
var returnValue =
                    seenTests.stream()
                            .map(
                                    test ->
                                            results.stream()
                                                    .filter(
                                                            result ->
                                                                    test.equals(result.testName()))
                                                    .findFirst()
                                                    .orElse(
                                                            new TestResult(
                                                                    test, TestStatus.UNRUN, false)))
                            .toList();
```

Ew. Ew ew ew. I do NOT like that at all. The tests are passing, so I took this opportunity for another pitstop in refactoring land. The concept we really want here is a map rather than a list, indexed by the name of the test, and then our functionality becomes a call to `getOrDefault()` on that map.

This ended up much cleaner but still a bit too indented for my liking giving how simple the actual behaviour is. I did a bit more refactoring here by adding an `unrun()` method to the `TestResult` record which wraps creation with the default fields.

Ahhhhh. Much nicer.