test(collator): Order, order!
This is one of those lovely little changes where the test is longer than the actual code change. I saw this one coming a few commits ago but now is finally the time to pay the piper. In Java, `HashSet` does not maintain insert ordering and instead orders based on the hashcode of the object. However, `TreeSet` _does_ do what I want.

This means that looping over the tests that the component has seen so far will maintain the order that is has ever seen the tests. If you run a single test once for the first time and never again then that test result is always going to be the first item in the list for the rest of time (or at least until the computer turns off).