test(collator): Order, order! (part 2)
I was a little bit naughty in the previous step. There is an extra part to ordering that I did not write a test for, so let's clean that up now.

Not only does the component maintain order across multiple runs, preserving a historical list of test order, but it also enforces that order within each run. This means that I can submit tests within a single run in the order D, C, B, A and they will be reported in the order A, B, C, D if that's the order they were originally seen.

Luckily, my code still passes the test but this means I went too far in my last change and wrote more code than I needed to pass the tests.