refactor(collator): Unification
It's largely feature complete at this point at least as far as the spec sets out. There's more I'd like to do with some of the `TestStatus` values; what does it mean exactly to disable a test? This could be a capability where the collator marks it with with that status and ignores any submitted results for it.

However, a bigger refactoring exposed itself when I took another step back. The component uses an ordered set to keep track of tests that it has already seen but uses an unordered map to keep track of the most recent batch of results.

Why not use an ordered map for both purposes? In this case, Java's `TreeMap`.

It sounds like a good idea but having separate components gives the component a way of knowing which test results have already been submitted and therefore which results are new.

Luckily for me, there is a neat way to handle this kind of scenario with Java's maps. Instead of using `.put()` and `.getOrDefault()`, I can use `.merge()`.

This means that where I was clearing the map of all entries at the end of each run, I can now use `.replaceAll()` to update the keys for every test and mark them as both unrun and not new.

If a new test is submitted, it is appended to the end of the map. If the test is not new, then we define the `.merge()` behaviour to be "take the status of the submitted test rather than unrun".

This tidies up the code significantly, removes the need for some boiler plate, and keeps all the test passing.

However, we can be cleaner with the representation of the domain which will be the final refactoring pitstop before we drive off into the sunset.