refactor(collator): Tidying up the domain
This is all working nicely but there's some oddness in our domain. Why does the code repeat the name of the test in both the key and the value of the results map?

I introduced a new concept called `StatusAndNew` and it largely does what it says on the tin; the trailing 2/3 of the original `TestResult` record. Now the code stores `StatusAndNew` against the test name instead of the whole object. This has two major implications.

First, it makes the reset logic slightly clearer as the `.replaceAll()` call does not depend on either the key _or_ the value in the map; it will always be `StatusAndNew(UNRUN, false)` so the record has a dedicated constructor for this case.

Second, it makes the _merge_ logic slightly clearer as "this test has already been seen" can be a method on the value rather than having to create an entirely new `TestResult`. This has the nice name of `.seen()` which describes exactly what the concept represents.

I do wonder if the design would have turned out different had the spec said `seen` instead of `isNew` (as they are inverses).

All in all, I'm quite happy with this. I didn't handle any of the error cases, such as what to do when `.add()` is called with a badly formed test name or null. There is a bit of type safety from the enum `TestStatus` but I could always wrap the "primitive" `String` in a `TestName` class that does all the right kind of validation.